Move `Type` pretty printer test into `core/src/pretty.rs`
